+++ Sources/Sentry/include/SentryPrivate.h
@@ -3,6 +3,7 @@
 #import "SentryBaggage.h"
 #import "SentryBaseIntegration.h"
 #import "SentryDispatchQueueWrapper.h"
+#import "SentryNSDataUtils.h"
 #import "SentryRandom.h"
 #import "SentrySdkInfo.h"
 #import "SentryStatsdClient.h"