+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features
 
+- Add Metrics API (#3791): Read our [docs](https://docs.sentry.io/platforms/apple/metrics/) to learn
+more about how to use the Metrics API.
 - Pre-main profiling data is now attached to the app start transaction (#3736)
 
 ### Fixes